Create User Input Concatenation
This Python program repeatedly prompts the user for input and stores the various inputs until the user types "\end".  At that point, the program prints a concatenated list of the various user inputs.  The program also capitalizes the first letter of the user's input and adds a question mark to questions and a period to all other sentence types.